feat(dynamodb): create initial protect-dynamo interface
