+++ examples/dynamo/package.json
@@ -22,6 +22,7 @@
     "@aws-sdk/lib-dynamodb": "^3.817.0",
     "@aws-sdk/util-dynamodb": "^3.817.0",
     "@cipherstash/protect": "workspace:*",
+    "@cipherstash/protect-dynamodb": "workspace:*",
     "pg": "^8.13.1"
   },
   "devDependencies": {
