+++ examples/dynamo/src/encrypted-partition-key.ts
@@ -1,12 +1,8 @@
 import { GetCommand, PutCommand } from '@aws-sdk/lib-dynamodb'
-import { createTable, docClient } from './common/dynamo'
-import {
-  decryptModel,
-  encryptModel,
-  makeSearchTerm,
-  users,
-} from './common/protect'
+import { createTable, docClient, dynamoClient } from './common/dynamo'
+import { users, protectClient } from './common/protect'
 import { log } from './common/log'
+import { protectDynamoDB } from '@cipherstash/protect-dynamodb'
 
 const tableName = 'UsersEncryptedPartitionKey'
 
@@ -31,6 +27,12 @@ const main = async () => {
     ],
   })
 
+  const protectDynamo = protectDynamoDB({
+    protectClient,
+    dynamoClient,
+    docClient,
+  })
+
   const user = {
     // `email` will be encrypted because it's included in the `users` protected table schema.
     email: 'abc@example.com',
@@ -38,7 +40,7 @@ const main = async () => {
     somePlaintextAttr: 'abc',
   }
 
-  const encryptResult = await encryptModel(user, users)
+  const encryptResult = await protectDynamo.encryptModel(user, users)
 
   log('encrypted item', encryptResult)
 
@@ -49,7 +51,11 @@ const main = async () => {
 
   await docClient.send(putCommand)
 
-  const searchTerm = await makeSearchTerm('abc@example.com', users.email, users)
+  const searchTerm = await protectDynamo.makeSearchTerm(
+    'abc@example.com',
+    users.email,
+    users,
+  )
 
   const getCommand = new GetCommand({
     TableName: tableName,
@@ -58,7 +64,10 @@ const main = async () => {
 
   const getResult = await docClient.send(getCommand)
 
-  const decryptedItem = await decryptModel<User>(getResult.Item, users)
+  const decryptedItem = await protectDynamo.decryptModel<User>(
+    getResult.Item,
+    users,
+  )
 
   log('decrypted item', decryptedItem)
 }
