+++ examples/dynamo/src/encrypted-sort-key.ts
@@ -1,12 +1,8 @@
 import { GetCommand, PutCommand } from '@aws-sdk/lib-dynamodb'
-import { createTable, docClient } from './common/dynamo'
-import {
-  decryptModel,
-  encryptModel,
-  makeSearchTerm,
-  users,
-} from './common/protect'
+import { createTable, docClient, dynamoClient } from './common/dynamo'
+import { users, protectClient } from './common/protect'
 import { log } from './common/log'
+import { protectDynamoDB } from '@cipherstash/protect-dynamodb'
 
 const tableName = 'UsersEncryptedSortKey'
 
@@ -40,6 +36,12 @@ const main = async () => {
     ],
   })
 
+  const protectDynamo = protectDynamoDB({
+    protectClient,
+    dynamoClient,
+    docClient,
+  })
+
   const user = {
     // `pk` won't be encrypted because it's not in the protected table schema.
     pk: 'user#1',
@@ -47,7 +49,7 @@ const main = async () => {
     email: 'abc@example.com',
   }
 
-  const encryptResult = await encryptModel(user, users)
+  const encryptResult = await protectDynamo.encryptModel(user, users)
 
   log('encrypted item', encryptResult)
 
@@ -58,7 +60,11 @@ const main = async () => {
 
   await docClient.send(putCommand)
 
-  const searchTerm = await makeSearchTerm('abc@example.com', users.email, users)
+  const searchTerm = await protectDynamo.makeSearchTerm(
+    'abc@example.com',
+    users.email,
+    users,
+  )
 
   const getCommand = new GetCommand({
     TableName: tableName,
@@ -67,7 +73,14 @@ const main = async () => {
 
   const getResult = await docClient.send(getCommand)
 
-  const decryptedItem = await decryptModel<User>(getResult.Item, users)
+  if (!getResult.Item) {
+    throw new Error('Item not found')
+  }
+
+  const decryptedItem = await protectDynamo.decryptModel<User>(
+    getResult.Item,
+    users,
+  )
 
   log('decrypted item', decryptedItem)
 }
