+++ examples/dynamo/src/simple.ts
@@ -1,7 +1,8 @@
 import { dynamoClient, docClient, createTable } from './common/dynamo'
 import { log } from './common/log'
-import { users, encryptModel, decryptModel } from './common/protect'
+import { users, protectClient } from './common/protect'
 import { GetCommand, PutCommand } from '@aws-sdk/lib-dynamodb'
+import { protectDynamoDB } from '@cipherstash/protect-dynamodb'
 
 const tableName = 'UsersSimple'
 
@@ -27,6 +28,12 @@ const main = async () => {
     ],
   })
 
+  const protectDynamo = protectDynamoDB({
+    protectClient,
+    dynamoClient,
+    docClient,
+  })
+
   const user = {
     // `pk` won't be encrypted because it's not included in the `users` protected table schema.
     pk: 'user#1',
@@ -34,7 +41,7 @@ const main = async () => {
     email: 'abc@example.com',
   }
 
-  const encryptResult = await encryptModel(user, users)
+  const encryptResult = await protectDynamo.encryptModel(user, users)
 
   log('encrypted item', encryptResult)
 
@@ -52,7 +59,10 @@ const main = async () => {
 
   const getResult = await docClient.send(getCommand)
 
-  const decryptedItem = await decryptModel<User>(getResult.Item, users)
+  const decryptedItem = await protectDynamo.decryptModel<User>(
+    getResult.Item,
+    users,
+  )
 
   log('decrypted item', decryptedItem)
 }
