+++ packages/protect-dynamodb/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@cipherstash/protect-dynamodb",
-	"version": "0.0.1",
+	"version": "0.0.0",
 	"description": "Protect.js DynamoDB Helpers",
 	"keywords": [
 		"dynamodb",
@@ -34,6 +34,8 @@
 		"release": "tsup"
 	},
 	"devDependencies": {
+		"@aws-sdk/client-dynamodb": "^3.817.0",
+		"@aws-sdk/lib-dynamodb": "^3.817.0",
 		"dotenv": "^16.4.7",
 		"tsup": "catalog:repo",
 		"tsx": "catalog:repo",
@@ -41,7 +43,7 @@
 		"vitest": "catalog:repo"
 	},
 	"peerDependencies": {
-		"@cipherstash/protect": "9.0.0"
+		"@cipherstash/protect": "workspace:*"
 	},
 	"publishConfig": {
 		"access": "public"
